First commit of PLS stopping criteria tests
Mostly this has a simple PLS and some functions for a variety of tests.

The data examples are not what we’d expect for PLS but (1) know the
structure of these data and (2) know what to expect from various
manipulations of these data

Next steps: (1) Better examples and more simulations required, and (2)
get split-half up and running (but this has many, many possible
criteria to test for).